apps/data-pipeline/dark_pool_processor.py
import asyncio
import logging
import json
import os
from datetime import datetime, timedelta
import requests
from bs4 import BeautifulSoup
import pandas as pd
from supabase import create_client, Client
from dotenv import load_dotenv

logger = logging.getLogger(__name__)
logging.basicConfig(level=logging.INFO, format='{"timestamp": "%(asctime)s", "level": "%(levelname)s", "component": "%(name)s", "message": "%(message)s", "metadata": %(metadata)s}')

load_dotenv()
SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")
supabase: Client = create_client(SUPABASE_URL, SUPABASE_KEY)

DARK_POOL_TABLE = "dark_pool_data"

class DarkPoolProcessor:
    def __init__(self):
        self.session = requests.Session()
        self.session.headers.update({
            "User-Agent": "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36"
        })
        # Note: In a production environment, you would use an actual API or data provider
        # for dark pool data. This is a simplified example.
        self.dark_pool_url = "https://example.com/dark-pool-data"  # Placeholder URL
    
    async def fetch_dark_pool_data(self, ticker):
        """
        Fetch dark pool data for a specific ticker.
        
        Note: This is a placeholder implementation. In a real-world scenario,
        you would subscribe to a financial data provider that offers dark pool data.
        """
        try:
            # Simulated data for demonstration purposes
            # In reality, you would make API calls to a data provider
            
            # Sample data structure
            current_date = datetime.now()
            data = {
                "ticker": ticker,
                "date": current_date.strftime("%Y-%m-%d"),
                "total_volume": 1000000,  # Example value
                "dark_pool_volume": 250000,  # Example value
                "dark_pool_percentage": 25.0,  # Example percentage
                "blocks": [
                    {
                        "time": (current_date - timedelta(hours=2)).strftime("%H:%M:%S"),
                        "price": 150.25,
                        "volume": 50000
                    },
                    {
                        "time": (current_date - timedelta(hours=1)).strftime("%H:%M:%S"),
                        "price": 151.30,
                        "volume": 75000
                    }
                ],
                "timestamp": current_date.isoformat()
            }
            
            return data
        except Exception as e:
            logger.error(f"Error fetching dark pool data for {ticker}: {e}", 
                        extra={"metadata": {"ticker": ticker}})
            return None
    
    def analyze_dark_pool_data(self, data):
        """
        Analyze dark pool data to identify patterns or significant activity.
        """
        if not data:
            return None
        
        try:
            ticker = data["ticker"]
            dark_pool_percentage = data["dark_pool_percentage"]
            dark_pool_volume = data["dark_pool_volume"]
            total_volume = data["total_volume"]
            
            # Calculate metrics
            significant_volume = dark_pool_volume > 100000  # Example threshold
            high_percentage = dark_pool_percentage > 20  # Example threshold
            
            # Check for block trades (large individual trades)
            large_blocks = [block for block in data["blocks"] if block["volume"] > 25000]  # Example threshold
            
            analysis = {
                "ticker": ticker,
                "date": data["date"],
                "dark_pool_percentage": dark_pool_percentage,
                "dark_pool_volume": dark_pool_volume,
                "total_volume": total_volume,
                "significant_volume": significant_volume,
                "high_percentage": high_percentage,
                "large_block_count": len(large_blocks),
                "largest_block": max([block["volume"] for block in data["blocks"]]) if data["blocks"] else 0,
                "analysis_timestamp": datetime.now().isoformat()
            }
            
            return analysis
        except Exception as e:
            logger.error(f"Error analyzing dark pool data: {e}", 
                        extra={"metadata": {"ticker": data.get("ticker", "unknown")}})
            return None
    
    def store_dark_pool_data(self, analysis):
        """
        Store dark pool analysis in Supabase database.
        """
        if not analysis:
            return
        
        try:
            # Check if entry for this ticker and date already exists
            existing = supabase.table(DARK_POOL_TABLE) \
                .select("id") \
                .eq("ticker", analysis["ticker"]) \
                .eq("date", analysis["date"]) \
                .execute()
            
            if not existing.data or len(existing.data) == 0:
                # Insert new analysis
                supabase.table(DARK_POOL_TABLE).insert(analysis).execute()
                logger.info(f"Inserted new dark pool analysis for {analysis['ticker']}", 
                          extra={"metadata": {"ticker": analysis["ticker"]}})
            else:
                # Update existing analysis
                supabase.table(DARK_POOL_TABLE) \
                    .update(analysis) \
                    .eq("ticker", analysis["ticker"]) \
                    .eq("date", analysis["date"]) \
                    .execute()
                logger.info(f"Updated dark pool analysis for {analysis['ticker']}", 
                          extra={"metadata": {"ticker": analysis["ticker"]}})
        except Exception as e:
            logger.error(f"Error storing dark pool analysis: {e}", 
                        extra={"metadata": {"ticker": analysis.get("ticker", "unknown")}})
    
    async def run(self, tickers):
        """
        Process dark pool data for a list of tickers.
        """
        for ticker in tickers:
            logger.info(f"Processing dark pool data for {ticker}", 
                       extra={"metadata": {"ticker": ticker}})
            
            data = await self.fetch_dark_pool_data(ticker)
            if data:
                analysis = self.analyze_dark_pool_data(data)
                if analysis:
                    self.store_dark_pool_data(analysis)
            
            # Avoid rate limiting
            await asyncio.sleep(1)
        
        logger.info(f"Completed dark pool processing for {len(tickers)} tickers", 
                   extra={"metadata": {"count": len(tickers)}})
    
    def close(self):
        """Close the session."""
        if self.session:
            self.session.close()

async def main():
    processor = DarkPoolProcessor()
    try:
        # Example list of tickers to process
        # In practice, you might get this from database or configuration
        tickers = ["AAPL", "MSFT", "NVDA", "TSLA", "AMZN"]
        await processor.run(tickers)
    finally:
        processor.close()

if __name__ == "__main__":
    loop = asyncio.get_event_loop()
    loop.run_until_complete(main()) 