apps/data-pipeline/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
supabase==1.0.3
python-dotenv==1.0.0
schedule==1.2.0
pandas==2.0.3
aiohttp==3.8.5
pandas-ta==0.3.14b0
flask==2.3.3
yfinance==0.2.28
pytest==7.4.0
asyncio==3.4.3 